added widget to separate files at reservation widget
