+++ lib/ui/widgets/Reservation/components/birthday_text_input_formatter.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/services.dart';
+
+class BirthTextInputFormatter extends TextInputFormatter {
+  @override
+  TextEditingValue formatEditUpdate(TextEditingValue oldValue, TextEditingValue newValue) {
+    if (oldValue.text.length >= newValue.text.length) {
+      return newValue;
+    }
+    var dateText = _addSeparator(newValue.text, '/');
+    return newValue.copyWith(text: dateText, selection: updateCursorPosition(dateText));
+  }
+
+  String _addSeparator(String value, String separator) {
+    value = value.replaceAll('/', '');
+    var newString = '';
+    for (int i = 0; i < value.length; i++) {
+      newString += value[i];
+      if (i == 1) {
+        newString += separator;
+      }
+      if (i == 3) {
+        newString += separator;
+      }
+    }
+    return newString;
+  }
+
+  TextSelection updateCursorPosition(String text) {
+    return TextSelection.fromPosition(TextPosition(offset: text.length));
+  }
+}
