+++ lib/ui/widgets/Reservation/components/example_mask.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
+
+class ExampleMask {
+  final TextEditingController textController = TextEditingController();
+  final MaskTextInputFormatter formatter;
+  final FormFieldValidator<String>? validator;
+  final String hint;
+  final TextInputType textInputType;
+
+  ExampleMask({required this.formatter, this.validator, required this.hint, required this.textInputType});
+}
