+++ lib/ui/widgets/Reservation/components/keys.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class Keys {
+  static final nameKey = GlobalKey<FormState>();
+  static final surnameKey = GlobalKey<FormState>();
+  static final birthdayKey = GlobalKey<FormState>();
+  static final citizenshipKey = GlobalKey<FormState>();
+  static final passportNumber = GlobalKey<FormState>();
+  static final passportValidityPeriod = GlobalKey<FormState>();
+}
