+++ lib/ui/widgets/Reservation/components/reservation_grey_data_text_widget.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:hotel_app/resources/app_colors.dart';
+
+class ReservationGreyDataText extends StatelessWidget {
+  const ReservationGreyDataText({
+    super.key,
+    required this.text,
+  });
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: 140,
+      child: Text(
+        text,
+        style: const TextStyle(fontSize: 16, fontWeight: FontWeight.w400, color: AppColors.greyText),
+      ),
+    );
+  }
+}
